adding protected route for patient
