+++ src/Scheduler/CustomModal.tsx
@@ -81,7 +81,7 @@ const CustomModal = ({
                 <ModalHeader>
                     {
                         `Appointment ${(typeof eventType.start === 'string' && eventType.start.length > 6) ?
-                            eventType.start.split('T')[1].substring(0, 5) : ''}-
+                            eventType.start.split('T')[1].substring(0, 5) : ''} -
                         ${(typeof eventType.end === 'string' && eventType.end.length > 6) ?
             eventType.end.split('T')[1].substring(0, 5) : ''}`
                     }
